Improved Movement

+++ src/main/kotlin/de/tho/neon/neon/io/business/GamePlayer.kt
@@ -6,7 +6,7 @@ import de.tho.neon.neon.io.model.Player
 data class GamePlayer(
     val id: Long,
     val user: Player,
-    var x: Long,
-    var y: Long,
-    var z: Long
+    var x: Double,
+    var y: Double,
+    var z: Double
 )
