+++ src/main/kotlin/de/tho/neon/neon/io/business/GamePosition.kt
@@ -2,7 +2,7 @@ package de.tho.neon.neon.io.business
 
 data class GamePosition(
     val id: Long,
-    val x: Long,
-    val y: Long,
-    val z: Long
+    val x: Double,
+    val y: Double,
+    val z: Double
 )
